Anonymize most callbacks and move into utils
